Add files via upload
We use this python script to send the situation of a traffic signal(in case all the lanes are heavy traffic) to our database. From there we will extract this information, display on our web app. We will also send a notification/alert to any user(saying that the particular intersection is full of traffic) if the user is in a certain radius with respect to the location of the traffic intersection mentioned before.