Add files via upload
This Code is Written in Arduino IDE . It was used to run our small scale traffic signal which served as a demo/ small scale implementation of our project(youtube link for the demo video is in the readme file in our repo). We used pyserial library to communicate between the arduino and our python script. The python script calculated the density of traffic, allotted time in seconds in a variable(say t1) for the traffic lane. The variable t1 was then passed onto the arduino, using the variable the arduino was able to regulate the amount of time the signal light stays green.